pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "pybind11>=2.10.0",
    "cmake>=3.15",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "openviking"
version = "0.1.6"
description = "An Agent-native context database: Data in, Context out"
readme = "README.md"
authors = [
    {name = "ByteDance", email = "noreply@bytedance.com"}
]
license = { text = "Apache-2.0" }
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.0.0",
    "typing-extensions>=4.5.0",
    "PyJWT[crypto]>=2.10.1",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "pdfplumber>=0.10.0",
    "readabilipy>=0.2.0",
    "markdownify>=0.11.0",
    "openai>=1.0.0",
    "requests>=2.28.0",
    "json-repair>=0.25.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
    "apscheduler>=3.11.0",
    "orjson>=3.11.4",
    "volcengine>=1.0.212",
    "volcengine-python-sdk[ark]>=5.0.3",
    "pyagfs>=1.4.0",
    "psutil>=7.2.1",
    "fastapi>=0.128.0",
    "uvicorn>=0.39.0",
    "werkzeug>=3.1.4",
    "pandas>=2.3.3",
    "xxhash>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/volcengine/openviking"
Documentation = "https://openviking.readthedocs.io"
Repository = "https://github.com/volcengine/openviking"
Issues = "https://github.com/volcengine/openviking/issues"

[project.scripts]
openviking-viewer = "openviking.tools.viewer:main"
openviking-github-app-review-bot = "openviking.tools.github_app_review_bot:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["openviking*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools.package-data]
openviking = ["prompts/templates/**/*.yaml"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = false
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=openviking --cov-report=term-missing"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "B006",  # Do not use mutable data structures for argument defaults
    "B904",  # Within an `except` clause, raise exceptions with `raise ... from err`
    "E741",  # Ambiguous variable name
    "E722",  # Do not use bare `except`
    "B027",  # empty method in an abstract base class
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
